+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - Added ERC721_deploy and ERC1155_deploy migration templates.
 - Added ERC1155_sealedSafeMint migration template.
 - Added ContractOwnership_transferOwnership migration template.
-- Added metadata resolvers migrations.
+- Added metadata resolvers migrations and templates.
 
 ## 2.0.0